Update Email.scala to keep API sound and make it simple to send text only emails
Hi,

While using this great plugin, I noticed an inconsistency in the Email object and case class declarations. The case class uses an Option[String]  for htmlText, the Object does not, using a simple String. While this is addressed by passing Some(htmlText) to the case class constructor, it makes it harder to send text only emails, since if I pass an empty string to the object, it will become Some("") instead of None. This way the email will be sent with an empty HTML message instead of just the text message...

Please consider the simple change proposed in this PR, at it makes it easier to send text only emails and improves consistency at the same time.

Thank you.

Cheers, Richard